move ImplementedInventory into util package
