+++ common/src/main/java/net/xstopho/resource_backpacks/util/BackpackInventory.java
@@ -4,7 +4,6 @@ import net.minecraft.core.NonNullList;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.component.ItemContainerContents;
-import net.xstopho.resource_backpacks.api.ImplementedInventory;
 
 public class BackpackInventory implements ImplementedInventory {
     private final ItemStack stack;
